client classes moved to communication folder
