+++ client/communication/ClientCommunicator.h
